Diet C codes
