+++ src/ext/Ppmd7.c
@@ -2,8 +2,6 @@
 2017-04-03 : Igor Pavlov : Public domain
 This code is based on PPMd var.H (2001): Dmitry Shkarin : Public domain */
 
-#include "Compiler.h"
-
 #include <string.h>
 
 #include "Ppmd7.h"
