+++ src/ext/Ppmd8.c
@@ -2,8 +2,6 @@
 2017-04-03 : Igor Pavlov : Public domain
 This code is based on PPMd var.I (2002): Dmitry Shkarin : Public domain */
 
-#include "Compiler.h"
-
 #include <string.h>
 
 #include "Ppmd8.h"
