+++ src/ext/Ppmd8Dec.c
@@ -4,8 +4,6 @@ This code is based on:
   PPMd var.I (2002): Dmitry Shkarin : Public domain
   Carryless rangecoder (1999): Dmitry Subbotin : Public domain */
 
-#include "Compiler.h"
-
 #include "Ppmd8.h"
 
 #define kTop (1 << 24)
